requirements.txt
#### DJANGO
django

#### LIBRARIES
pyyaml
pytest
certifi
pytest-cov
pytest-django
psycopg2-binary
django-livereload-server